pyproject.toml
[project]
name = "restaurant-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.0",
    "streamlit>=1.46.0",
    "python-dotenv>=1.0.1",
]
